requirements.txt
apispec==6.4.0
apispec-webframeworks==0.5.2
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.2
packaging==23.2
psycopg==3.1.17
PyJWT==2.8.0
python-dotenv==1.0.0
PyYAML==6.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
